Extract RealtimeChannel MessageFilter-related code to new class
In preparation for #1397 (making MessageFilter functionality
tree-shakable).
